requirements.txt
numpy
torch
gymnasium[box2d]
tqdm
tensorboard
tensorboardX
imageio
Pillow